Now the app gets data from a restaurant's JSON through API

+++ app/src/main/java/com/alexiusdev/depeat/datamodels/Product.java
@@ -3,9 +3,9 @@ package com.alexiusdev.depeat.datamodels;
 public class Product {
     private String name;
     private int quantity;
-    private float price;
+    private double price;
 
-    public Product(String name, int quantity, float price) {
+    public Product(String name, int quantity, double price) {
         this.name = name;
         this.quantity = quantity;
         this.price = price;
@@ -27,7 +27,7 @@ public class Product {
         this.quantity = quantity;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
